Update PythonProject.py
This code fetches COVID-19 data for all countries,checks if the request is successful then parses the JSON response and sorts the countries based on their deaths in descending order. It then prints the names and deaths of the top 10 countries. In case of a failed request, it prints an error message along with the HTTP status code.